+++ src/script/jswd.js
@@ -2,10 +2,10 @@ const _ = require('underscore');
 
 let moduleExports = _.extend(
     require('jswd/module'),
+    require('jswd/processor'),
     require('jswd/register'),
     require('jswd/symbol'),
     require('jswd/virtual'),
-
 );
 
 moduleExports.physical = require('jswd/physical');
